requirements.txt
asgiref==3.2.10
astroid==2.4.2
cffi==1.14.2
Django==2.2.13
django-filter==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
isort==5.5.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mysqlclient==2.0.1
pycparser==2.20
PyJWT==1.6.4
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
